requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
diskcache==5.6.3
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-text-splitters==0.0.1
langserve==0.0.51
langsmith==0.1.31
llama-cpp-python==0.2.57
MarkupSafe==2.1.5
marshmallow==3.21.1
motor==3.4.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.9.15
packaging==23.2
pydantic==2.6.4
pydantic-core==2.16.3
pymongo==4.6.3
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sniffio==1.3.1
SQLAlchemy==2.0.28
sse-starlette==2.0.0
starlette==0.36.3
tenacity==8.2.3
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.39.1
typing-extensions==4.10.0
typing-inspect==0.9.0
urllib3==2.2.1
uvicorn==0.29.0
yarl==1.9.4
